ex2 add prop

+++ src/ru/voothi/lesson5/phw/Animal.java
@@ -2,14 +2,14 @@ package ru.voothi.lesson5.phw;
 
 public class Animal {
     private int length;
-    private int hight;
+    private float hight;
 
-    public void run(int length) {
+    public void run(float length) {
     }
 
-    public void swim(int length) {
+    public void swim(float length) {
     }
 
-    public void jump(int hight) {
+    public void jump(float hight) {
     }
 }
